+++ PostHog/PostHogSDK.swift
@@ -47,7 +47,7 @@ let maxRetryDelay = 30.0
     private var flagCallReported: [String: [Any?]] = .init()
     private(set) var remoteConfig: PostHogRemoteConfig?
     private var context: PostHogContext?
-    private static var apiKeys = Set<String>()
+    private static var projectTokens = Set<String>()
     private var installedIntegrations: [PostHogIntegration] = []
     let sessionManager = PostHogSessionManager()
     let onEventCaptured = PostHogMulticastCallback<PostHogEvent>()
@@ -61,9 +61,9 @@ let maxRetryDelay = 30.0
 
     // nonisolated(unsafe) is introduced in Swift 5.10
     #if swift(>=5.10)
-        @objc public nonisolated(unsafe) static let shared: PostHogSDK = .init(PostHogConfig(apiKey: ""))
+        @objc public nonisolated(unsafe) static let shared: PostHogSDK = .init(PostHogConfig(projectToken: ""))
     #else
-        @objc public static let shared: PostHogSDK = .init(PostHogConfig(apiKey: ""))
+        @objc public static let shared: PostHogSDK = .init(PostHogConfig(projectToken: ""))
     #endif
 
     deinit {
@@ -91,10 +91,10 @@ let maxRetryDelay = 30.0
                 return
             }
 
-            if PostHogSDK.apiKeys.contains(config.apiKey) {
-                hedgeLog("API Key: \(config.apiKey) already has a PostHog instance.")
+            if PostHogSDK.projectTokens.contains(config.projectToken) {
+                hedgeLog("Project token: \(config.projectToken) already has a PostHog instance.")
             } else {
-                PostHogSDK.apiKeys.insert(config.apiKey)
+                PostHogSDK.projectTokens.insert(config.projectToken)
             }
 
             enabled = true
@@ -1767,7 +1767,7 @@ let maxRetryDelay = 30.0
 
         setupLock.withLock {
             enabled = false
-            PostHogSDK.apiKeys.remove(config.apiKey)
+            PostHogSDK.projectTokens.remove(config.projectToken)
 
             queue?.stop()
             replayQueue?.stop()
@@ -1776,7 +1776,7 @@ let maxRetryDelay = 30.0
             replayQueue = nil
             config.storageManager?.reset(keepAnonymousId: config.reuseAnonymousId)
             config.storageManager = nil
-            config = PostHogConfig(apiKey: "")
+            config = PostHogConfig(projectToken: "")
             remoteConfig = nil
             storage = nil
             #if !os(watchOS)
@@ -1855,11 +1855,10 @@ let maxRetryDelay = 30.0
                 return hedgeLog("Could not start recording. Session replay feature flag is disabled.")
             }
 
-            let sessionId = resumeCurrent
+            guard (resumeCurrent
                 ? sessionManager.getSessionId()
-                : sessionManager.getNextSessionId()
-
-            guard let sessionId else {
+                : sessionManager.getNextSessionId()) != nil
+            else {
                 return hedgeLog("Could not start recording. Missing session id.")
             }